Add postinstall script. (#3)
Adds a postinstall script that will run when npm install is executed. The script copies the files in the node_modules/pspdfkit/dist/modern/pspdfkit-lib folder in staticresources and modern/pspdfkit.js, filtering the files that are not necessary for the Salesforce integration so as to avoid surpassing the file limit imposed by the Salesforce environment.